+++ pkg/boxesimpl/boxesdocfactory.go
@@ -72,6 +72,7 @@ func initBoxFormat(f *boxes.Format) boxes.BoxFormat {
 	boxMargin := types.GlobalMinBoxMargin
 	var fixedHeight, fixedWidth, cornerRadius *int
 	var widthOfParent *bool
+	var renderType *types.BoxRenderType
 	if f != nil {
 		initFontFormatAnchorInCase(f.FontCaption)
 		initFontFormatAnchorInCase(f.FontText1)
@@ -96,6 +97,14 @@ func initBoxFormat(f *boxes.Format) boxes.BoxFormat {
 		if f.VerticalTxt != nil {
 			verticalTxt = *f.VerticalTxt
 		}
+
+		if f.RenderType != nil {
+			renderType = f.RenderType
+		} else {
+			rt := types.BoxRenderTypeRectangle
+			renderType = &rt
+		}
+
 		fixedHeight = f.FixedHeight
 		fixedWidth = f.FixedWidth
 		cornerRadius = f.CornerRadius
@@ -117,6 +126,7 @@ func initBoxFormat(f *boxes.Format) boxes.BoxFormat {
 		WidthOfParent:     widthOfParent,
 		VerticalTxt:       verticalTxt,
 		CornerRadius:      cornerRadius,
+		RenderType:        renderType,
 	}
 }
 
@@ -356,6 +366,14 @@ func initLayoutElement(l *boxes.Layout, doc *boxes.BoxesDocument, b *boxes.Boxes
 	if l.Id != "" {
 		newParents = append(newParents, l.Id)
 	}
+	dontBlockConPaths := l.DontBlockConPaths
+	format := adjustFormatBasedOnVariations(l, b, f, doc)
+	if format != nil && format.RenderType != nil {
+		if *format.RenderType != types.BoxRenderTypeRectangle {
+			t := true
+			dontBlockConPaths = &t
+		}
+	}
 	elem := boxes.LayoutElement{
 		Id:                l.Id,
 		Caption:           l.Caption,
@@ -366,8 +384,8 @@ func initLayoutElement(l *boxes.Layout, doc *boxes.BoxesDocument, b *boxes.Boxes
 		Image:             initImage(l, b.Images),
 		Vertical:          initLayoutElemContainer(l.Vertical, doc, b, newParents),
 		Horizontal:        initLayoutElemContainer(l.Horizontal, doc, b, newParents),
-		Format:            adjustFormatBasedOnVariations(l, b, f, doc),
-		DontBlockConPaths: l.DontBlockConPaths,
+		Format:            format,
+		DontBlockConPaths: dontBlockConPaths,
 		DataLink:          l.DataLink,
 		Connections:       initConnections(l.Connections, doc.Formats),
 		ParentIds:         parents,