+++ pkg/types/boxes/boxes_mixins.gen.go
@@ -17,13 +17,19 @@ import (
 type ProcessStep struct {
     // title, that's appended to the original layout title
 Caption string `yaml:"caption"`
+    // dictionary for layout mixins. key of the dictionary is the caption of the box that will take the additional content
+LayoutMixins map[string]LayoutMixin `yaml:"layoutMixins,omitempty"`
+    // array for box mixins. It's useful in cases where the dictionary reference to boxes to mix in is limiting
+BoxMixins []BoxMixin `yaml:"boxMixins,omitempty"`
     // dictionary of connection objects
 Connections map[string]ConnectionCont `yaml:"connections,omitempty"`
     // dictionary of comment objects, this comment will applied on layout objects and replace existing comments there
 Comments map[string]types.Comment `yaml:"comments,omitempty"`
     // dictionary of tag array, the additional tags will be applied on the existing layout and can be used for instance to define display formats
-Tags map[string]Tags `yaml:"tags,omitempty"`
+Tags map[string][]string `yaml:"tags,omitempty"`
 Overlays []Overlay `yaml:"overlays,omitempty"`
+    // Be careful it can mess up the stype of the picture! This allows to overwrite global formats in an available step. It can be helpful in cases this step wants to highlight things, that are visual pressed back by existing formats
+Formats map[string]Format `yaml:"formats,omitempty"`
     // Set of formats that overwrites the style of boxes, if specific conditions are met
 FormatVariations *FormatVariations `yaml:"formatVariations,omitempty"`
 }
@@ -36,6 +42,20 @@ func CopyProcessStep(src *ProcessStep) *ProcessStep {
     var ret ProcessStep
 
     ret.Caption = src.Caption
+
+    if src.LayoutMixins != nil {
+        ret.LayoutMixins = make(map[string]LayoutMixin, len(src.LayoutMixins))
+        for k, v := range src.LayoutMixins {
+            ret.LayoutMixins[k] = *CopyLayoutMixin(&v)
+        }
+    }
+
+    if src.BoxMixins != nil {
+        ret.BoxMixins = make([]BoxMixin, len(src.BoxMixins))
+        for i, v := range src.BoxMixins {
+            ret.BoxMixins[i] = *CopyBoxMixin(&v)
+        }
+    }
 
     if src.Connections != nil {
         ret.Connections = make(map[string]ConnectionCont, len(src.Connections))
@@ -52,9 +72,9 @@ func CopyProcessStep(src *ProcessStep) *ProcessStep {
     }
 
     if src.Tags != nil {
-        ret.Tags = make(map[string]Tags, len(src.Tags))
+        ret.Tags = make(map[string][]string, len(src.Tags))
         for k, v := range src.Tags {
-            ret.Tags[k] = *CopyTags(&v)
+            ret.Tags[k] = v
         }
     }
 
@@ -65,6 +85,13 @@ func CopyProcessStep(src *ProcessStep) *ProcessStep {
         }
     }
 
+    if src.Formats != nil {
+        ret.Formats = make(map[string]Format, len(src.Formats))
+        for k, v := range src.Formats {
+            ret.Formats[k] = *CopyFormat(&v)
+        }
+    }
+
     ret.FormatVariations = CopyFormatVariations(src.FormatVariations)
 return &ret
 }
@@ -72,10 +99,202 @@ return &ret
 
 func NewProcessStep() *ProcessStep {
     var ret ProcessStep
+    ret.LayoutMixins = make(map[string]LayoutMixin, 0)
+    ret.BoxMixins = make([]BoxMixin, 0)
     ret.Connections = make(map[string]ConnectionCont, 0)
     ret.Comments = make(map[string]types.Comment, 0)
-    ret.Tags = make(map[string]Tags, 0)
+    ret.Tags = make(map[string][]string, 0)
     ret.Overlays = make([]Overlay, 0)
+    ret.Formats = make(map[string]Format, 0)
+    return &ret
+}
+
+// definition of a box to be mixed in
+type BoxMixin struct {
+    // unique identifier of that entry
+Id string `yaml:"id"`
+    // Some kind of the main text
+Caption string `yaml:"caption"`
+    // First additional text
+Text1 string `yaml:"text1"`
+    // Second additional text
+Text2 string `yaml:"text2"`
+    // additional comment, that can be then included in the created graphic
+Comment *types.Comment `yaml:"comment,omitempty"`
+    // Reference to an image that should be displayed, needs to be declared in the global image section
+Image *string `yaml:"image,omitempty"`
+    // in case the picture is rendered with given expanded IDs, and maxDepth, then if this flag is true, the box is still displayed expanded
+Expand bool `yaml:"expand"`
+    // If set, then the content for 'vertical' attrib is loaded from an external file
+ExtVertical *string `yaml:"extVertical,omitempty"`
+Vertical []Layout `yaml:"vertical,omitempty"`
+    // If set, then the content for 'horizontal' attrib is loaded from an external file
+ExtHorizontal *string `yaml:"extHorizontal,omitempty"`
+Horizontal []Layout `yaml:"horizontal,omitempty"`
+    // Tags to annotate the box, tags are used to format and filter
+Tags []string `yaml:"tags,omitempty"`
+    // List of connections to other boxes
+Connections []Connection `yaml:"connections,omitempty"`
+    // reference to the format to use for this box
+Format *string `yaml:"format,omitempty"`
+    // if that is set then connections can run through the box, as long as they don't cross the text
+DontBlockConPaths *bool `yaml:"dontBlockConPaths,omitempty"`
+    // Optional link to a source, related to this element. This can be used for instance for on-click handlers in a UI or simply as documentation.
+DataLink *string `yaml:"dataLink,omitempty"`
+    // is only set by while the layout is processed, don't set it in the definition
+HiddenComments bool `yaml:"hiddenComments"`
+    // either ID or caption of the box where this mixin is placed before or after
+Reference string `yaml:"reference"`
+    // triggers the new mixin elemente to put either left or above the object with the given reference
+PutBefore *bool `yaml:"putBefore,omitempty"`
+    // triggers the new mixin elemente to put either right or below the object with the given reference
+PutAfter *bool `yaml:"putAfter,omitempty"`
+    // use this field if you want to wrap existing horizontal or vertical containers in it's own box.
+WrapSubElems *SubsWrapperObj `yaml:"wrapSubElems,omitempty"`
+}
+
+
+func CopyBoxMixin(src *BoxMixin) *BoxMixin {
+    if src == nil {
+        return nil
+    }
+    var ret BoxMixin
+
+    ret.Id = src.Id
+
+    ret.Caption = src.Caption
+
+    ret.Text1 = src.Text1
+
+    ret.Text2 = src.Text2
+
+    ret.Comment = types.CopyComment(src.Comment)
+
+    if src.Image != nil {
+        v := *src.Image
+        ret.Image = &v
+    }
+
+    ret.Expand = src.Expand
+
+    if src.ExtVertical != nil {
+        v := *src.ExtVertical
+        ret.ExtVertical = &v
+    }
+
+    if src.Vertical != nil {
+        ret.Vertical = make([]Layout, len(src.Vertical))
+        for i, v := range src.Vertical {
+            ret.Vertical[i] = *CopyLayout(&v)
+        }
+    }
+
+    if src.ExtHorizontal != nil {
+        v := *src.ExtHorizontal
+        ret.ExtHorizontal = &v
+    }
+
+    if src.Horizontal != nil {
+        ret.Horizontal = make([]Layout, len(src.Horizontal))
+        for i, v := range src.Horizontal {
+            ret.Horizontal[i] = *CopyLayout(&v)
+        }
+    }
+
+    if src.Tags != nil {
+        ret.Tags = make([]string, len(src.Tags))
+        for i, v := range src.Tags {
+            ret.Tags[i] = v
+        }
+    }
+
+    if src.Connections != nil {
+        ret.Connections = make([]Connection, len(src.Connections))
+        for i, v := range src.Connections {
+            ret.Connections[i] = *CopyConnection(&v)
+        }
+    }
+
+    if src.Format != nil {
+        v := *src.Format
+        ret.Format = &v
+    }
+
+    if src.DontBlockConPaths != nil {
+        v := *src.DontBlockConPaths
+        ret.DontBlockConPaths = &v
+    }
+
+    if src.DataLink != nil {
+        v := *src.DataLink
+        ret.DataLink = &v
+    }
+
+    ret.HiddenComments = src.HiddenComments
+
+    ret.Reference = src.Reference
+
+    if src.PutBefore != nil {
+        v := *src.PutBefore
+        ret.PutBefore = &v
+    }
+
+    if src.PutAfter != nil {
+        v := *src.PutAfter
+        ret.PutAfter = &v
+    }
+
+    ret.WrapSubElems = CopySubsWrapperObj(src.WrapSubElems)
+return &ret
+}
+
+
+func NewBoxMixin() *BoxMixin {
+    var ret BoxMixin
+    ret.Vertical = make([]Layout, 0)
+    ret.Horizontal = make([]Layout, 0)
+    ret.Tags = make([]string, 0)
+    ret.Connections = make([]Connection, 0)
+    return &ret
+}
+
+type BoxMixinBase struct {
+    // either ID or caption of the box where this mixin is placed before or after
+Reference string `yaml:"reference"`
+    // triggers the new mixin elemente to put either left or above the object with the given reference
+PutBefore *bool `yaml:"putBefore,omitempty"`
+    // triggers the new mixin elemente to put either right or below the object with the given reference
+PutAfter *bool `yaml:"putAfter,omitempty"`
+    // use this field if you want to wrap existing horizontal or vertical containers in it's own box.
+WrapSubElems *SubsWrapperObj `yaml:"wrapSubElems,omitempty"`
+}
+
+
+func CopyBoxMixinBase(src *BoxMixinBase) *BoxMixinBase {
+    if src == nil {
+        return nil
+    }
+    var ret BoxMixinBase
+
+    ret.Reference = src.Reference
+
+    if src.PutBefore != nil {
+        v := *src.PutBefore
+        ret.PutBefore = &v
+    }
+
+    if src.PutAfter != nil {
+        v := *src.PutAfter
+        ret.PutAfter = &v
+    }
+
+    ret.WrapSubElems = CopySubsWrapperObj(src.WrapSubElems)
+return &ret
+}
+
+
+func NewBoxMixinBase() *BoxMixinBase {
+    var ret BoxMixinBase
     return &ret
 }
 
@@ -106,33 +325,6 @@ func NewConnectionCont() *ConnectionCont {
     return &ret
 }
 
-type Tags struct {
-Tags []string `yaml:"tags,omitempty"`
-}
-
-
-func CopyTags(src *Tags) *Tags {
-    if src == nil {
-        return nil
-    }
-    var ret Tags
-
-    if src.Tags != nil {
-        ret.Tags = make([]string, len(src.Tags))
-        for i, v := range src.Tags {
-            ret.Tags[i] = v
-        }
-    }
-return &ret
-}
-
-
-func NewTags() *Tags {
-    var ret Tags
-    ret.Tags = make([]string, 0)
-    return &ret
-}
-
 // Model to inject additional things in a boxes layout definition
 type BoxesFileMixings struct {
     // optional title, that's appended to the original layout title
@@ -143,6 +335,8 @@ Version *string `yaml:"version,omitempty"`
 Legend *Legend `yaml:"legend,omitempty"`
     // dictionary for layout mixins. key of the dictionary is the caption of the box that will take the additional content
 LayoutMixins map[string]LayoutMixin `yaml:"layoutMixins,omitempty"`
+    // array for box mixins. It's useful in cases where the dictionary reference to boxes to mix in is limiting
+BoxMixins []BoxMixin `yaml:"boxMixins,omitempty"`
     // dictionary of connection objects
 Connections map[string]ConnectionCont `yaml:"connections,omitempty"`
 Formats map[string]Format `yaml:"formats,omitempty"`
@@ -151,7 +345,7 @@ FormatVariations *FormatVariations `yaml:"formatVariations,omitempty"`
     // dictionary of comment objects, this comment will applied on layout objects and replace existing comments there
 Comments map[string]types.Comment `yaml:"comments,omitempty"`
     // dictionary of tag array, the additional tags will be applied on the existing layout and can be used for instance to define display formats
-Tags map[string]Tags `yaml:"tags,omitempty"`
+Tags map[string][]string `yaml:"tags,omitempty"`
     // optional map of images used in the generated graphic
 Images map[string]types.ImageDef `yaml:"images,omitempty"`
 Overlays []Overlay `yaml:"overlays,omitempty"`
@@ -185,6 +379,13 @@ func CopyBoxesFileMixings(src *BoxesFileMixings) *BoxesFileMixings {
         }
     }
 
+    if src.BoxMixins != nil {
+        ret.BoxMixins = make([]BoxMixin, len(src.BoxMixins))
+        for i, v := range src.BoxMixins {
+            ret.BoxMixins[i] = *CopyBoxMixin(&v)
+        }
+    }
+
     if src.Connections != nil {
         ret.Connections = make(map[string]ConnectionCont, len(src.Connections))
         for k, v := range src.Connections {
@@ -209,9 +410,9 @@ func CopyBoxesFileMixings(src *BoxesFileMixings) *BoxesFileMixings {
     }
 
     if src.Tags != nil {
-        ret.Tags = make(map[string]Tags, len(src.Tags))
+        ret.Tags = make(map[string][]string, len(src.Tags))
         for k, v := range src.Tags {
-            ret.Tags[k] = *CopyTags(&v)
+            ret.Tags[k] = v
         }
     }
 
@@ -242,10 +443,11 @@ return &ret
 func NewBoxesFileMixings() *BoxesFileMixings {
     var ret BoxesFileMixings
     ret.LayoutMixins = make(map[string]LayoutMixin, 0)
+    ret.BoxMixins = make([]BoxMixin, 0)
     ret.Connections = make(map[string]ConnectionCont, 0)
     ret.Formats = make(map[string]Format, 0)
     ret.Comments = make(map[string]types.Comment, 0)
-    ret.Tags = make(map[string]Tags, 0)
+    ret.Tags = make(map[string][]string, 0)
     ret.Images = make(map[string]types.ImageDef, 0)
     ret.Overlays = make([]Overlay, 0)
     ret.Steps = make([]ProcessStep, 0)